--- src/main/java/io/pravega/connectors/flink/sink/PravegaWriter.java
@@ -1,214 +1,0 @@
-/**
- * Copyright Pravega Authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.pravega.connectors.flink.sink;
-
-import io.pravega.client.ClientConfig;
-import io.pravega.client.stream.Serializer;
-import io.pravega.client.stream.Stream;
-import io.pravega.client.stream.TxnFailedException;
-import io.pravega.connectors.flink.PravegaEventRouter;
-import io.pravega.connectors.flink.PravegaWriterMode;
-import org.apache.flink.annotation.VisibleForTesting;
-import org.apache.flink.api.common.serialization.SerializationSchema;
-import org.apache.flink.api.connector.sink.Sink;
-import org.apache.flink.api.connector.sink.SinkWriter;
-import org.apache.flink.metrics.Gauge;
-import org.apache.flink.metrics.MetricGroup;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import javax.annotation.Nullable;
-import java.io.IOException;
-import java.nio.ByteBuffer;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.stream.Collectors;
-
-/**
- * This is a adapter for initialize, write, and prepareCommit stages
- * of the underlying {@link FlinkPravegaInternalWriter}.<p>
- * See also {@link PravegaCommitter} for the commit stage.
- *
- * @param <T> The type of the event to be written.
- */
-public class PravegaWriter<T> implements SinkWriter<T, PravegaTransactionState, Void> {
-    private static final Logger LOG = LoggerFactory.getLogger(PravegaWriter.class);
-    private static final long serialVersionUID = 1L;
-    private static final String PRAVEGA_WRITER_METRICS_GROUP = "PravegaWriter";
-    private static final String SCOPED_STREAM_METRICS_GAUGE = "stream";
-
-    // --------- configurations for creating a FlinkPravegaInternalWriter ---------
-    protected final ClientConfig clientConfig;
-    protected final Stream stream;
-    protected final long txnLeaseRenewalPeriod;
-    protected final SerializationSchema<T> serializationSchema;
-    // The router used to partition events within a stream, can be null for random routing
-    @Nullable
-    protected final PravegaEventRouter<T> eventRouter;
-    // --------- configurations for creating a FlinkPravegaInternalWriter ---------
-
-    // The sink's mode of operation. This is used to provide different guarantees for the written events.
-    protected final PravegaWriterMode writerMode;
-
-    // The writer we use for create and begin a transaction.
-    // However, the transaction is committed in PravegaCommitter via a re-created FlinkPravegaInternalWriter.
-    protected FlinkPravegaInternalWriter<T> currentWriter;
-
-    // Place where we hold writers for different checkpoints.
-    private final List<FlinkPravegaInternalWriter<T>> writers = new ArrayList<>();
-
-    /**
-     * A Pravega Writer that implements {@link SinkWriter}.
-     *
-     * @param context               The sink initContext which we used only for metrics.
-     * @param enableMetrics         Flag to indicate whether metrics needs to be enabled or not.
-     * @param clientConfig          The Pravega client configuration.
-     * @param stream                The destination stream.
-     * @param txnLeaseRenewalPeriod Transaction lease renewal period in milliseconds.
-     * @param writerMode            The Pravega writer mode.
-     * @param serializationSchema   The implementation for serializing every event into pravega's storage format.
-     * @param eventRouter           The implementation to extract the partition key from the event.
-     */
-    public PravegaWriter(Sink.InitContext context,
-                         boolean enableMetrics,
-                         ClientConfig clientConfig,
-                         Stream stream,
-                         long txnLeaseRenewalPeriod,
-                         final PravegaWriterMode writerMode,
-                         SerializationSchema<T> serializationSchema,
-                         @Nullable PravegaEventRouter<T> eventRouter) {
-        this.writerMode = writerMode;
-
-        this.clientConfig = clientConfig;
-        this.stream = stream;
-        this.txnLeaseRenewalPeriod = txnLeaseRenewalPeriod;
-        this.serializationSchema = serializationSchema;
-        this.eventRouter = eventRouter;
-
-        // the (transactional) pravega writer is initialized
-        // in FlinkPravegaInternalWriter#createInternalWriter
-        this.currentWriter = createFlinkPravegaInternalWriter();
-
-        if (this.writerMode == PravegaWriterMode.EXACTLY_ONCE) {
-            this.currentWriter.beginTransaction();
-
-            writers.add(currentWriter);
-        }
-
-        if (enableMetrics) {
-            MetricGroup pravegaWriterMetricGroup = context.metricGroup().addGroup(PRAVEGA_WRITER_METRICS_GROUP);
-            pravegaWriterMetricGroup.gauge(SCOPED_STREAM_METRICS_GAUGE, new StreamNameGauge(stream.getScopedName()));
-        }
-    }
-
-    @Override
-    public void write(T element, Context context) throws IOException {
-        try {
-            currentWriter.write(element);
-        } catch (TxnFailedException e) {
-            throw new IOException(e);
-        }
-    }
-
-    @Override
-    public List<PravegaTransactionState> prepareCommit(boolean flush) throws IOException {
-        final List<PravegaTransactionState> transactionStates;
-        try {
-            if (writerMode != PravegaWriterMode.BEST_EFFORT || flush) {
-                currentWriter.flushAndVerify();
-            }
-
-            switch (writerMode) {
-                case EXACTLY_ONCE:
-                    transactionStates = writers.stream().map(PravegaTransactionState::of).collect(Collectors.toList());
-                    writers.clear();
-
-                    currentWriter = createFlinkPravegaInternalWriter();
-                    currentWriter.beginTransaction();
-                    writers.add(currentWriter);
-
-                    break;
-                case ATLEAST_ONCE:
-                case BEST_EFFORT:
-                    transactionStates = new ArrayList<>();
-                    break;
-                default:
-                    throw new UnsupportedOperationException("Not implemented writer mode");
-            }
-        } catch (InterruptedException | TxnFailedException e) {
-            throw new IOException("", e);
-        }
-        LOG.info("Committing {} committables, final commit={}.", transactionStates, flush);
-        return transactionStates;
-    }
-
-    @Override
-    public void close() throws Exception {
-        LOG.info("Calling close from PravegaWriter");
-        currentWriter.close();
-    }
-
-    /**
-     * Wrap the SerializationSchema to a Pravega compatible Serializer.
-     * @param <T> The type of the event.
-     */
-    @VisibleForTesting
-    static final class FlinkSerializer<T> implements Serializer<T> {
-
-        private final SerializationSchema<T> serializationSchema;
-
-        FlinkSerializer(SerializationSchema<T> serializationSchema) {
-            this.serializationSchema = serializationSchema;
-        }
-
-        @Override
-        public ByteBuffer serialize(T value) {
-            return ByteBuffer.wrap(serializationSchema.serialize(value));
-        }
-
-        @Override
-        public T deserialize(ByteBuffer serializedValue) {
-            throw new IllegalStateException("deserialize() called within a serializer");
-        }
-    }
-
-    /**
-     * Gauge for getting the fully qualified stream name information.
-     */
-    private static class StreamNameGauge implements Gauge<String> {
-
-        final String stream;
-
-        public StreamNameGauge(String stream) {
-            this.stream = stream;
-        }
-
-        @Override
-        public String getValue() {
-            return stream;
-        }
-    }
-
-    /**
-     * This initialization method is extracted only fot the testing sake.
-     * @return The internal writer
-     */
-    @VisibleForTesting
-    protected FlinkPravegaInternalWriter<T> createFlinkPravegaInternalWriter() {
-        return new FlinkPravegaInternalWriter<>(clientConfig, stream,
-                txnLeaseRenewalPeriod, writerMode, serializationSchema, eventRouter);
-    }
-}
